*** empty log message ***

+++ trunk/DataSetTools/operator/SpectrometerTofToEnergy.java
@@ -115,7 +115,7 @@ public class SpectrometerTofToEnergy extends    DataSetOperator
                                      // get the energy scale parameters 
     float min_E = ( (Float)(getParameter(0).getValue()) ).floatValue();
     float max_E = ( (Float)(getParameter(1).getValue()) ).floatValue();
-    int   num_E = ( (Float)(getParameter(2).getValue()) ).intValue() + 1;
+    int   num_E = ( (Integer)(getParameter(2).getValue()) ).intValue() + 1;
 
                                      // validate energy bounds
     if ( min_E > max_E )             // swap bounds to be in proper order
